% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png.R
\name{as.png.data.frame}
\alias{as.png.data.frame}
\title{Coerce to PNG from Data Frame.}
\usage{
\method{as.png}{data.frame}(x, replace = TRUE, ...)
}
\arguments{
\item{x}{data.frame}

\item{replace}{whether to delete the intermediate PDF if successful}

\item{...}{passed to \code{\link{as.pdf.data.frame}} and to \code{\link{ghostconvert}}}
}
\value{
path to the file created
}
\description{
Coerces to PNG from 'data.frame'.
}
\examples{
\dontrun{
browseURL(as.png(head(Theoph), dir = tempdir(), gs_cmd = 'mgs'))
}
}
\seealso{
\code{\link{as.pdf.data.frame}}

\code{\link{ghostconvert}}
}
