% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{breaks}
\alias{breaks}
\title{Calculate Row and Column Breaks}
\usage{
breaks(x, ...)
}
\arguments{
\item{x}{vector}

\item{...}{ignored}
}
\value{
integer
}
\description{
Calculates row and column breaks by finding first elements among repeats.
}
\seealso{
\code{\link{as.tabular}}
}
