% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{command}
\alias{command}
\title{Format a Latex Command}
\usage{
command(x, options = NULL, args = NULL, depth = 0)
}
\arguments{
\item{x}{length one character}

\item{options}{vector or list}

\item{args}{vector or list}

\item{depth}{integer}
}
\value{
character
}
\description{
Format a latex command.  \code{x} is formated as a latex command, with the options (possibly named) inserted in square brackets, and the arguments each enclosed in curly braces.  \code{depth} spaces are added to the left end of the string.
}
\examples{
command('caption',options='short',args='long')

}
\seealso{
wrap

as.ltable.data.frame
}
