% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reserve.R
\name{reserve.character}
\alias{reserve.character}
\title{Escape LaTeX Reserved Characters for Character Objects}
\usage{
\method{reserve}{character}(
  x,
  target = c("#", "$", "\%", "^", "&", "_", "{", "}", "~", "\\\\"),
  replacement = c("\\\\#", "\\\\$", "\\\\\%", "\\\\^{}", "\\\\&", "\\\\_", "\\\\{",
    "\\\\}", "\\\\~{}", "\\\\textbackslash{}"),
  ...
)
}
\arguments{
\item{x}{character}

\item{target}{sequence to find (regular expression unless fixed is false)}

\item{replacement}{sequence to use}

\item{...}{passed arguments}
}
\description{
Escapes LaTeX reserved characters as suggested by \url{https://en.wikibooks.org/wiki/LaTeX/Basics#Reserved_Characters}.
}
\seealso{
Other reserve: 
\code{\link{reserve.data.frame}()},
\code{\link{reserve.default}()},
\code{\link{reserve.factor}()},
\code{\link{reserve.list}()},
\code{\link{reserve}()}
}
\concept{reserve}
