% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reserve.R
\name{reserve.data.frame}
\alias{reserve.data.frame}
\title{Escape LaTeX Reserved Characters in Data Frames}
\usage{
\method{reserve}{data.frame}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Applies \code{reserve} to the columns of a data.frame.
}
\examples{
foo <- c('#','$\%^','&_{','}~\\\\')
bar <- data.frame(
 stringsAsFactors = FALSE,
 a = as.numeric(factor(foo)),
 b = foo,
 c = factor(foo)
)

reserve(bar)
# as.pdf(bar, wider = 10)
# as.pdf(bar, target = '#') # fails

}
\seealso{
Other reserve: \code{\link{reserve.character}},
  \code{\link{reserve.default}},
  \code{\link{reserve.factor}}, \code{\link{reserve.list}},
  \code{\link{reserve}}
}
