% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reserve.R
\name{reserve.default}
\alias{reserve.default}
\title{Escape LaTeX Reserved Characters Using Default Method}
\usage{
\method{reserve}{default}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
The default method returns its argument.
}
\seealso{
Other reserve: \code{\link{reserve.character}},
  \code{\link{reserve.data.frame}},
  \code{\link{reserve.factor}}, \code{\link{reserve.list}},
  \code{\link{reserve}}
}
