% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reserve.R
\name{reserve.factor}
\alias{reserve.factor}
\title{Escape LaTeX Reserved Characters for Factor Objects}
\usage{
\method{reserve}{factor}(x, ...)
}
\arguments{
\item{x}{factor}

\item{...}{passed arguments}
}
\description{
Coerces to character and escapes reserved characters.
}
\seealso{
Other reserve: 
\code{\link{reserve.character}()},
\code{\link{reserve.data.frame}()},
\code{\link{reserve.default}()},
\code{\link{reserve.list}()},
\code{\link{reserve}()}
}
\concept{reserve}
