% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reserve.R
\name{reserve.list}
\alias{reserve.list}
\title{Escape LaTeX Reserved Characters in Lists}
\usage{
\method{reserve}{list}(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Applies \code{reserve} to the elements of a list.
}
\seealso{
Other reserve: 
\code{\link{reserve.character}()},
\code{\link{reserve.data.frame}()},
\code{\link{reserve.default}()},
\code{\link{reserve.factor}()},
\code{\link{reserve}()}
}
\concept{reserve}
