% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reserve.R
\name{reserve}
\alias{reserve}
\title{Escape LaTeX Reserved Characters}
\usage{
reserve(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}
}
\description{
Escapes LaTeX reserved characters.  Generic, with default, character, and data.frame methods.
}
\seealso{
Other reserve: \code{\link{reserve.character}},
  \code{\link{reserve.data.frame}},
  \code{\link{reserve.default}},
  \code{\link{reserve.factor}}, \code{\link{reserve.list}}
}
