% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{wrap}
\alias{wrap}
\title{Wrap Text in a Latex Environment}
\usage{
wrap(x, environment, options = NULL, args = NULL, depth = 0)
}
\arguments{
\item{x}{character}

\item{environment}{name of environment}

\item{options}{list or vector}

\item{args}{list or vector}

\item{depth}{integer (extra spaces on the left)}
}
\value{
character
}
\description{
Wrap text in a latex environment. \code{x} is wrapped in the specified environment, with options and arguments formatted appropriately.
}
\examples{
wrap('Hello','center')
}
\seealso{
\code{\link{command}}

\code{\link{as.ltable.data.frame}}
}
