% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{as.pdf}
\alias{as.pdf}
\alias{as.pdf.document}
\alias{as.pdf.character}
\title{Coerce to PDF}
\usage{
as.pdf(x, ...)

\method{as.pdf}{document}(x, stem = "latexpdf-doc", dir = ".",
  clean = TRUE, ...)

\method{as.pdf}{character}(x, stem, ...)
}
\arguments{
\item{x}{object}

\item{...}{passed arguments}

\item{stem}{the stem of a file name (no extension)}

\item{dir}{output directory}

\item{clean}{whether to delete system files after pdf creation}
}
\value{
the output file path (invisible)
}
\description{
Coerces to PDF. Generic, with methods for character, document, and data.frame.

Coerces to PDF from document. Makes a system call to 'pdflatex'. Extra arguments ignored.

Coerces character to document and then to PDF. Extra arguments passed to \code{\link{as.document.character}} and \code{\link{as.pdf.document}}.
}
\section{Methods (by class)}{
\itemize{
\item \code{document}: document method

\item \code{character}: character method
}}

\seealso{
\code{\link{as.pdf.character}}

\code{\link{as.pdf.document}}

\code{\link{as.pdf.data.frame}}

\code{\link{as.png}}

\code{\link{tex2pdf}}

\code{\link{viewtex}}
}
